Tidied up unit tests and added documentation
