+++ test/requirements_normalize_version_test.py
@@ -4,6 +4,10 @@ from lambda_package.requirements import normalize_version
 
 
 class RequirementsNormalizeVersionTests(unittest.TestCase):
+    """
+    Unit tests for the `requirements.normalize_version` function
+    """
+
     def test_when_call_with_valid_version_then_no_change(self):
         result = normalize_version("5.2")
         self.assertEqual("5.2", result)
